refactor(#4712): de-giant idle recap card lifecycle
Move the D6 card compose/post/persist cluster without behavior changes so idle_recap.rs falls below the giant-file threshold.
