+++ scripts/giant_file_registry.toml
@@ -782,22 +782,10 @@ owner = "discord-relay"
 deadline = "2026-08-31"
 decompose_issue = "#4712"
 
-# NOTE: src/services/discord/idle_recap.rs previously had an [[entry]] here; #3479
-# extracted the scrollback-capture/Haiku-summarizer and token-context-display
-# clusters into `idle_recap/scrollback.rs` and `idle_recap/context_display.rs`,
-# dropping its production LoC below the giant threshold. Its entry (and ratchet
-# baseline) were removed to avoid a ghost registration.
-
-[[entry]]
-# Re-crossed the 1000-prod-LoC threshold in #3809 when idle recap gained the
-# deterministic relay-integrity probe wiring, suspect-only diagnostic action,
-# and suggested-reply action plan. Keep any further relay-probe/reporting
-# growth in `idle_recap/` child modules and decompose the recap interaction
-# orchestration once the card-action contract stabilizes.
-file = "src/services/discord/idle_recap.rs"
-owner = "discord-relay"
-deadline = "2026-08-31"
-decompose_issue = "#4712"
+# RETIRED 2026-07-24 (#4712 D6): idle_recap.rs dropped below the production
+# giant threshold after the card compose/post/delete/persist cluster moved
+# verbatim to `idle_recap/card.rs`. Remove the [[entry]] to avoid a ghost
+# registration; keep future card lifecycle growth in the child module.
 
 # NOTE: src/services/discord/turn_bridge/terminal_delivery.rs previously had an
 # [[entry]] here; the #3028 splitter fix dropped its production LoC below the
